+++ libgsh/Cargo.toml
@@ -5,11 +5,9 @@ edition = "2021"
 
 [dependencies]
 shared = { path = "../shared" }
-rustls = { version = "0.23.26", features = [
-	"ring",
-	"std",
-], default-features = false }
 rcgen = "0.13.2"
 log = "0.4.27"
 thiserror = "2.0.12"
 tokio = { version = "1.44.2", features = ["full"] }
+tokio-rustls = { version = "0.26.2", features = ["ring"] }
+async-trait = "0.1.88"
