+++ libgsh/src/async/server.rs
@@ -1,15 +1,13 @@
-use std::{
-    net::{TcpListener, TcpStream},
-    sync::Arc,
-};
-
-use rustls::{ServerConfig, ServerConnection, StreamOwned};
-use shared::{protocol::client_hello, MessageCodec};
-
 use super::service::AsyncService;
+use crate::Result;
+use shared::{protocol::client_hello, r#async::AsyncMessageCodec};
+use std::sync::Arc;
+use tokio::net::{TcpListener, TcpStream};
+use tokio_rustls::rustls::ServerConfig;
+use tokio_rustls::{server::TlsStream, TlsAcceptor};
 
 const DEFAULT_PORT: u16 = 1122;
-pub type Messages = MessageCodec<StreamOwned<ServerConnection, TcpStream>>;
+pub type AsyncMessages = AsyncMessageCodec<TlsStream<TcpStream>>;
 
 /// An async server that handles client connections and manages the application service implementing the `AsyncService` trait.
 /// The server listens for incoming connections and spawns a new tasks for each client connection.\
@@ -29,7 +27,10 @@ pub struct AsyncServer<ServiceT: AsyncService> {
     config: ServerConfig,
 }
 
-impl<ServiceT: AsyncService> AsyncServer<ServiceT> {
+impl<ServiceT: AsyncService> AsyncServer<ServiceT>
+where
+    ServiceT: Send + Sync + 'static,
+{
     /// Creates a new `AsyncServer` instance with the provided server configuration.\
     /// The `ServerConfig` is used to configure the TLS settings for the server.
     pub fn new(config: ServerConfig) -> Self {
@@ -41,14 +42,15 @@ impl<ServiceT: AsyncService> AsyncServer<ServiceT> {
 
     /// Starts the server and listens for incoming connections on the default port (1122).\
     /// This method blocks until the server is stopped or an error occurs.
-    pub fn serve(self) -> Result<(), Box<dyn std::error::Error>> {
-        self.serve_port(DEFAULT_PORT)
+    pub async fn serve(self) -> Result<()> {
+        self.serve_port(DEFAULT_PORT).await
     }
 
     /// Starts the server and listens for incoming connections on the specified port.\
     /// This method blocks until the server is stopped or an error occurs.
-    pub fn serve_port(self, port: u16) -> Result<(), Box<dyn std::error::Error>> {
-        let listener = TcpListener::bind(format!("[::]:{}", port))?;
+    pub async fn serve_port(self, port: u16) -> Result<()> {
+        let listener = TcpListener::bind(format!("[::]:{}", port)).await?;
+        let tls_acceptor = TlsAcceptor::from(Arc::new(self.config.clone()));
         let service_fullname = std::any::type_name::<ServiceT>();
         let service_name = service_fullname
             .split("::")
@@ -59,32 +61,40 @@ impl<ServiceT: AsyncService> AsyncServer<ServiceT> {
             service_name,
             listener.local_addr()?
         );
-        while let Ok((mut stream, addr)) = listener.accept() {
-            let mut conn = ServerConnection::new(Arc::new(self.config.clone()))?;
-            conn.complete_io(&mut stream)?;
-            let tls_stream = StreamOwned::new(conn, stream);
-            let mut messages = Messages::new(tls_stream);
-            let client = shared::handshake_server(
-                &mut messages,
-                &[shared::PROTOCOL_VERSION],
-                ServiceT::server_hello(),
-            )?;
-            let os: client_hello::Os = client.os.try_into().unwrap_or(client_hello::Os::Unknown);
-            let monitors = client.monitors.len();
-            println!(
-                "+ Client connected running {:?} {} with {} monitor(s) on {}",
-                os,
-                client.os_version,
-                monitors,
-                addr.port(),
-            );
-            std::thread::spawn(move || {
-                if let Err(e) = ServiceT::new().main(messages) {
+        loop {
+            let (stream, addr) = listener.accept().await?;
+            let tls_acceptor = tls_acceptor.clone();
+            tokio::spawn(async move {
+                let tls_stream = tls_acceptor.accept(stream).await.unwrap();
+                let messages = AsyncMessages::new(tls_stream);
+                if let Err(e) = Self::handle_client(messages, addr).await {
                     log::error!("Service error {}: {}", addr, e);
                 }
                 println!("- Client disconnected from {}", addr);
             });
         }
+    }
+
+    /// Handles a client connection.\
+    /// This function performs the TLS handshake and starts the service's main event loop.\
+    async fn handle_client(mut messages: AsyncMessages, addr: std::net::SocketAddr) -> Result<()> {
+        let client = shared::r#async::handshake_server(
+            &mut messages,
+            &[shared::PROTOCOL_VERSION],
+            ServiceT::server_hello(),
+        )
+        .await?;
+        let os: client_hello::Os = client.os.try_into().unwrap_or(client_hello::Os::Unknown);
+        let monitors = client.monitors.len();
+        log::info!(
+            "+ Client connected running {:?} {} with {} monitor(s) on {}",
+            os,
+            client.os_version,
+            monitors,
+            addr.port()
+        );
+        let service = ServiceT::new();
+        service.main(messages).await?;
         Ok(())
     }
 }
