+++ shared/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 [dependencies]
 prost = "0.13.5"
 os_info = "3.10.0"
+tokio = { version = "1.44.2", features = ["full"] }
 
 [build-dependencies]
 prost-build = "0.13.5"
